Remove unused prop called onDownloadSelected
